Theme the address and rename boxes, move All drives to the top, fix a format specifier
The child EDIT controls showed as a stark white box against the dark
chrome. WM_CTLCOLOREDIT now paints them in the panel colours, so the
address bar and the rename field read as fields in this window rather
than holes cut out of it.

The All drives card moves to the top of the drive list, above the
per-drive cards: it reads as the whole machine first, stays in reach as
drives come and go, and never scrolls off.

And a format specifier a test used to print the parent reference stayed
%u after that reference widened to 64 bits, which code scanning flagged.
It is a test's diagnostic line, not shipping code and not a
vulnerability, but passing a 64-bit value to %u is undefined, so it is
now %llu with the matching cast.

+++ tests/test_ntfs.cpp
@@ -331,8 +331,10 @@ static void TestRecordParsing() {
     CHECK(info.parent == 42, "parent reference");
     CHECK(info.size == 512, "resident $DATA size");
     CHECK(!info.isExtension, "not an extension record");
-    std::printf("    parsed '%ls' parent=%u size=%llu\n", info.name.c_str(),
-                info.parent, static_cast<unsigned long long>(info.size));
+    // parent is 64-bit now (the full $FILE_NAME reference); print it as one.
+    std::printf("    parsed '%ls' parent=%llu size=%llu\n", info.name.c_str(),
+                static_cast<unsigned long long>(info.parent),
+                static_cast<unsigned long long>(info.size));
 
     rec = MakeRecord(L"Users", 5, 0, true);
     ApplyFixups(rec.data(), rec.size(), 512);
